Prepocessed Data from MIMIC III
The data from MIMIC III was processed using Big Query.
The data has been processed to include variables to prove our hypothesis.
The hypothesis is that the variation of BP can be a predivtor of mortality in "moderately ill patients".
The criteria for moderately ill patients was suggested by physician after looking at the distribution of the data.